updated migrations

+++ Modules/BooksModule/Database/migrations/2024_04_30_181741_create_books_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('books', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('num_of_read_pages');
+            $table->unsignedBigInteger('num_of_read_pages');
             // $table->timestamps();
 
             $table->index('num_of_read_pages');
